Move attentional-focus search to here
