+++ opencog/attentionbank/AttentionalFocusCB.h
@@ -1,7 +1,9 @@
 /*
- * BindLinkAPI.h
+ * AttentionalFocusCB.h
  *
- * Copyright (C) 2014 Linas Vepstas <linasvepstas@gmail.com>
+ * Copyright (C) 2014 Misgana Bayetta
+ *
+ * Author: Misgana Bayetta <misgana.bayetta@gmail.com>  July 2014
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License v3 as
@@ -18,22 +20,27 @@
  * Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
+#ifndef _ATTENTIONAL_FOCUS_CB_H
+#define _ATTENTIONAL_FOCUS_CB_H
 
-#ifndef _OPENCOG_BINDLINK_API_H
-#define _OPENCOG_BINDLINK_API_H
-
-#include <opencog/atoms/base/Handle.h>
-#include <opencog/truthvalue/TruthValue.h>
+#include <opencog/query/DefaultPatternMatchCB.h>
 
 namespace opencog {
 
-class AtomSpace;
+class AttentionalFocusCB: public virtual DefaultPatternMatchCB
+{
+public:
+	AttentionalFocusCB(AtomSpace*);
 
-Handle bindlink(AtomSpace*, const Handle&, size_t max_results=SIZE_MAX);
-TruthValuePtr satisfaction_link(AtomSpace*, const Handle&);
-Handle satisfying_set(AtomSpace*, const Handle&, size_t max_results=SIZE_MAX);
-Handle recognize(AtomSpace*, const Handle&);
+	// Only match nodes if they are in the attentional focus
+	bool node_match(const Handle&, const Handle&);
 
-} // namespace opencog
+	// Only match links if they are in the attentional focus
+	bool link_match(const PatternTermPtr&, const Handle&);
 
-#endif // _OPENCOG_BINDLINK_API_H
+	// Only get incoming sets that are in the attentional focus
+	IncomingSet get_incoming_set(const Handle&);
+};
+
+} //namespace opencog
+#endif /* _ATTENTIONALFOCUSCB_H */
